Write check results to file and show progress
This redirects the output of the constraint check to a file 'out.csv'.
The output file name will change as soon as we have the command line
option to specify an input file. Additionally, progress of the check
is indicated to the command line using '.', '-' and '+' characters as
well as an occasional update on how many items have been processed
overall (every 100 items).

Also, a few example input files have been added in the data/ directory.